Moved files to public-api
