+++ services/public-api/internal/models/models.go
@@ -1,4 +1,4 @@
-package transformations
+package models
 
 import (
 	"time"
